+++ lib/models/bingo_hall_model.dart
@@ -19,6 +19,7 @@ abstract class BingoHallModel with _$BingoHallModel {
     String? city,
     String? state,
     String? zipCode,
+    String? unitNumber,
     String? phone,
     String? websiteUrl,
     String? description,